+++ role.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Role(Enum):
+    SPINE = "spine"
+    LEAF = "leaf"
